首页编写完毕

+++ idf-admin/src/main/java/com/idofast/admin/controller/vo/request/ProxyInfoUpdateVo.java
@@ -26,7 +26,7 @@ public class ProxyInfoUpdateVo
     private Long speed;
 
     @ApiModelProperty("用户总流量")
-    private Integer totalData;
+    private Long totalData;
 
 
     @ApiModelProperty("下一个结算日期；时间戳格式")
